requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.0
django-crispy-forms==1.11.2
django-heroku==0.3.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
Pillow==8.2.0
psycopg2==2.8.6
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.5.67
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5